+++ Magic/CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fix sniper rifle damage not undoing
  - The /mskills menu will now work with a player's class skills
  - Add quick_cast and skill_limit properties for classes to control /mskills behavior
+ - Add auto undo to the Lightning spell
 
 ## 7.3